refactor: move commission and related entities to pkg/entity
