+++ pkg/entity/comparable_number.go
