+++ pkg/internal/storage/database/scopes/for_gorm_gen.go
@@ -2,6 +2,7 @@ package scopes
 
 import (
 	"fmt"
+
 	"github.com/4chain-ag/go-wallet-toolbox/pkg/internal/storage/queryopts"
 	"gorm.io/gen"
 	"gorm.io/gen/field"
