+++ pkg/storage/crud/numeric_cond.go
@@ -1,7 +1,7 @@
 package crud
 
 import (
-	"github.com/4chain-ag/go-wallet-toolbox/pkg/internal/storage/entity"
+	"github.com/4chain-ag/go-wallet-toolbox/pkg/entity"
 	"github.com/go-softwarelab/common/pkg/types"
 )
